feat: useToggleOption 추가
